# Optimized the performance of schema change, improved by 1.42-27 times