--- a/be/src/storage/convert_helper.h
+++ b/be/src/storage/convert_helper.h
@@ -41,8 +41,7 @@ public:
     MaterializeTypeConverter() = default;
     ~MaterializeTypeConverter() = default;
 
-    virtual Status convert_materialized(ColumnPtr src_col, ColumnPtr dst_col, TypeInfo* src_type,
-                                        const TabletColumn& ref_column) const = 0;
+    virtual Status convert_materialized(ColumnPtr src_col, ColumnPtr dst_col, TypeInfo* src_type) const = 0;
 };
 
 const MaterializeTypeConverter* get_materialized_converter(FieldType from_type, MaterializeType to_type);