… be/src/storage/schema_change.cpp